* added nodejs18.x as latest

* added just nodejs

* version bump

* Update package.json

* Allowing for storing github token in settings file instead of env var

* Allowing for storing github token in settings file instead of env var

* new feature:  command that generates patterns using ChatGPT

* new feature:  command that generates patterns using ChatGPT

* new feature:  command that generates patterns using ChatGPT

* new feature:  command that generates patterns using ChatGPT

* make GPT model selectable

* cdk support in 'generate' command

* support for GPT generated state machines and lambda functions

* version bump

---------

+++ src/commands/generate/generate.js
@@ -4,10 +4,16 @@ const { Configuration, OpenAIApi } = require("openai");
 const settingsUtil = require("../../shared/settingsUtil");
 const fs = require("fs-extra");
 var Spinner = require('cli-spinner').Spinner;
-
+const baseFile = require("../../shared/baseFile.json");
 var spinner = new Spinner('Waiting for ChatGPT... %s');
 spinner.setSpinnerString('|/-\\');
 async function run(cmd) {
+  console.log("*** Note: This is an experimental feature and depends on the ChatGPT API. Make sure you review the output carefully before using it in production ***")
+  const output = cmd.output.toLowerCase();
+  if (output !== "sam" && !cmd.outputFile) {
+    console.log(`You need to specify an output file with --output-file`);
+    return;
+  }
   const apiKey = settingsUtil.getConfigSource().openaiApiKey;
   if (!apiKey) {
     console.log(
@@ -15,27 +21,54 @@ async function run(cmd) {
     );
     return;
   }
-  if (!fs.existsSync(cmd.template)) {
-    console.log(
-      `Can't find ${cmd.template}. Use -t option to specify template filename`
-    );
-    const create = await inputUtil.prompt(`Create ${cmd.template}?`);
-    if (create) {
-      fs.writeFileSync(cmd.template, parser.stringify("yaml", baseFile));
-    } else {
-      return;
-    }
-  }
 
-  const ownTemplate = parser.parse("own", fs.readFileSync(cmd.template));
-  ownTemplate.Resources = ownTemplate.Resources || {};
+  let ownTemplate;
+  if (output === "sam") {
+    if (!fs.existsSync(cmd.template)) {
+      console.log(
+        `Can't find ${cmd.template}. Use -t option to specify template filename`
+      );
+      const create = await inputUtil.prompt(`Create ${cmd.template}?`);
+      if (create) {
+        fs.writeFileSync(cmd.template, parser.stringify("yaml", baseFile));
+      } else {
+        return;
+      }
+    }
+
+    ownTemplate = parser.parse("own", fs.readFileSync(cmd.template));
+    ownTemplate.Resources = ownTemplate.Resources || {};
+  }
 
   const configuration = new Configuration({
     apiKey,
   });
   const openai = new OpenAIApi(configuration);
+  let outputString;
+  let language;
+  switch (output.substring(0, 3)) {
+    case "sam":
+      outputString = "SAM JSON";
+      break;
+    case "cdk":
+      outputString = "TypeScript CDK";
+      break;
+    case "lam":
+      outputString = "Lambda";
+      language = output.split("-")[1];
+      if (language) {
+        outputString += ` in ${language}`;
+      }
 
-  const prompt = `Generate this in AWS SAM JSON: ${cmd.query}`;
+      break;
+    case "asl":
+      outputString = "StepFunctions ASL YAML";
+      break;
+    default:
+      console.log(`Invalid output format ${output}. Valid values are 'SAM', 'CDK', 'lambda-<language>' or 'ASL'`);
+      return;
+  }
+  const prompt = `Generate this in AWS ${outputString}: ${cmd.query}. Only return code.`;
 
   const openAiRequest = {
     model: cmd.model,
@@ -51,26 +84,54 @@ async function run(cmd) {
   spinner.start();
   const response = await openai.createChatCompletion(openAiRequest);
   spinner.stop();
-  const text = response.data.choices[0].message.content;
+  let text = response.data.choices[0].message.content;
   // get the first JSON object in the text
   let obj;
-  try {
-    obj = JSON.parse(text.replace(/\n/g, '').replace(/```/g, '').match(/{.*}/)[0]);
-  } catch (e) {
+  if (output === "sam") {
     try {
-      obj = parser.parse("yaml", text);
+      obj = JSON.parse(text.replace(/\n/g, '').replace(/```/g, '').match(/{.*}/)[0]);
     } catch (e) {
-      console.log(`Couldn't parse the output from ChatGPT. Try again. The output was: \n${text}\n\nThe error was: ${e}`);
-      return;
+      try {
+        obj = parser.parse("yaml", text);
+      } catch (e) {
+        console.log(`Couldn't parse the output from ChatGPT. Try again. The output was: \n${text}\n\nThe error was: ${e}`);
+        return;
+      }
     }
+    if (obj.Resources) {
+      obj = obj.Resources;
+    }
+
+    console.log(`\n\nGenerated the following resources:\n\n${parser.stringify("yaml", obj)}`);
+  } else {
+    console.log(text);
+    //check if text has a row starting with ``` followed by text. If so, remove that text
+    const match = text.match(/```.+\n/);
+    if (match && match[0]) {
+      text = text.replace(match[0], '```\n');
+    }
+    if (text.match(/```/g)) {
+      text = text.replace(/\n/g, "¶").match(/```.*```/)[0].split("¶").join('\n').replace(/```/g, '');
+    }
+    console.log(`\n\nGenerated the following ${language} code:\n\n${text}`);
   }
-  if (obj.Resources) {
-    obj = obj.Resources;
+  const cont = await inputUtil.prompt(output === "sam" ? `Add to template?` : `Add to ${cmd.outputFile}?`);
+  if (!cont) {
+    return;
   }
 
-  console.log(`\n\nGenerated the following resources:\n\n${parser.stringify("yaml", obj)}`);
-  const cont = await inputUtil.prompt(`Add to template?`);
-  if (!cont) {
+  if (output !== "sam") {
+    if (!cmd.outputFile) {
+      console.log(`You need to specify an output file with --output-file`);
+      return;
+    }
+    if (fs.existsSync(cmd.outputFile)) {
+      const cont = await inputUtil.prompt(`Overwrite ${cmd.outputFile}?`);
+      if (!cont) {
+        return;
+      }
+    }
+    fs.writeFileSync(cmd.outputFile, text);
     return;
   }
